Factor out YAML parsing code to table/meta.py
